@@ -10,45 +10,50 @@
     "node": ">=12.x"
   },
   "dependencies": {
-    "babel-plugin-styled-components": "~1.10.6",
-    "babel-preset-react": "~6.24.1",
-    "gatsby": "~2.18.17",
-    "gatsby-link": "~2.2.27",
-    "gatsby-plugin-dark-mode": "~1.1.0",
-    "gatsby-plugin-feed": "~2.3.25",
-    "gatsby-plugin-manifest": "~2.2.34",
-    "gatsby-plugin-offline": "~3.0.30",
-    "gatsby-plugin-react-helmet": "~3.1.18",
-    "gatsby-plugin-sharp": "~2.3.10",
-    "gatsby-plugin-styled-components": "~3.1.16",
-    "gatsby-plugin-typography": "~2.3.20",
-    "gatsby-remark-copy-linked-files": "~2.1.33",
-    "gatsby-remark-images": "~3.1.39",
-    "gatsby-remark-prismjs": "~3.3.28",
-    "gatsby-remark-responsive-iframe": "~2.2.30",
-    "gatsby-remark-smartypants": "~2.1.19",
-    "gatsby-source-filesystem": "~2.1.43",
-    "gatsby-transformer-remark": "~2.6.45",
-    "gatsby-transformer-sharp": "~2.3.9",
-    "lodash": "~4.17.15",
-    "prismjs": "~1.17.1",
-    "react": "~16.12.0",
-    "react-dom": "~16.12.0",
-    "react-helmet": "~5.2.1",
-    "react-icons": "~3.8.0",
-    "react-typography": "~0.16.19",
-    "styled-components": "~4.4.1",
-    "typeface-merriweather": "~0.0.72",
-    "typeface-montserrat": "~0.0.75",
+    "babel-plugin-styled-components": "^1.11.1",
+    "babel-preset-react": "^6.24.1",
+    "dotenv": "^8.2.0",
+    "gatsby": "^2.24.85",
+    "gatsby-link": "~2.4.16",
+    "gatsby-plugin-feed": "^2.5.19",
+    "gatsby-plugin-manifest": "^2.4.35",
+    "gatsby-plugin-offline": "^3.2.37",
+    "gatsby-plugin-react-helmet": "^3.3.14",
+    "gatsby-plugin-sharp": "^2.6.43",
+    "gatsby-plugin-styled-components": "^3.3.14",
+    "gatsby-plugin-typography": "^2.5.13",
+    "gatsby-remark-copy-linked-files": "^2.3.19",
+    "gatsby-remark-images": "^3.3.39",
+    "gatsby-remark-prismjs": "^3.5.16",
+    "gatsby-remark-responsive-iframe": "^2.4.17",
+    "gatsby-remark-smartypants": "^2.3.13",
+    "gatsby-source-filesystem": "^2.3.35",
+    "gatsby-transformer-remark": "^2.8.46",
+    "gatsby-transformer-sharp": "^2.5.19",
+    "prismjs": "^1.22.0",
+    "react": "~16.13.1",
+    "react-dom": "~16.13.1",
+    "react-helmet": "^6.1.0",
+    "react-icons": "^3.11.0",
+    "react-typography": "^0.16.19",
+    "styled-components": "^5.2.0",
+    "typeface-merriweather": "^1.1.13",
+    "typeface-montserrat": "^1.1.13",
     "typography": "~0.16.19",
-    "typography-theme-wordpress-2016": "~0.16.19"
+    "typography-theme-wordpress-2016": "^0.16.19"
   },
   "devDependencies": {
-    "babel-preset-gatsby": "~0.2.26",
-    "eslint": "~6.8.0",
-    "eslint-plugin-react": "~7.17.0",
-    "gh-pages": "~2.1.1",
-    "prettier": "~1.19.1"
+    "babel-eslint": "^10.1.0",
+    "cross-env": "^7.0.2",
+    "babel-preset-gatsby": "^0.5.14",
+    "eslint": "^7.12.0",
+    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-react-hooks": "^4.1.2",
+    "esm": "^3.2.25",
+    "prettier": "^2.1.2"
   },
   "homepage": "https://github.com/Cabeda/jecabeda-blog",
   "keywords": [
