@@ -1,68 +1,61 @@
-const _ = require('lodash')
-const Promise = require('bluebird')
-const path = require('path')
-const { createFilePath } = require('gatsby-source-filesystem')
+import path from "path";
+import { createFilePath } from "gatsby-source-filesystem";
 
-exports.createPages = ({ graphql, actions }) => {
-  const { createPage } = actions
+async function createBlogPosts({ graphql, actions }) {
+  // 1. Get a template file for this page
+  const blogTemplate = path.resolve("./src/templates/blog-post.js");
 
-  return new Promise((resolve, reject) => {
-    const blogPost = path.resolve('./src/templates/blog-post.js')
-    resolve(
-      graphql(
-        `
-          {
-            allMarkdownRemark(sort: { fields: [frontmatter___date], order: DESC }, limit: 1000) {
-              edges {
-                node {
-                  fields {
-                    slug
-                  }
-                  frontmatter {
-                    title
-                  }
-                }
-              }
+  // 2. Query all posts
+  const { data } = await graphql(`
+    query {
+      allMarkdownRemark(
+        sort: { fields: [frontmatter___date], order: DESC }
+        limit: 1000
+      ) {
+        edges {
+          node {
+            fields {
+              slug
+            }
+            frontmatter {
+              title
             }
           }
-        `
-      ).then(result => {
-        if (result.errors) {
-          console.log(result.errors)
-          reject(result.errors)
         }
+      }
+    }
+  `);
 
-        // Create blog posts pages.
-        const posts = result.data.allMarkdownRemark.edges;
+  const posts = data.allMarkdownRemark.edges;
 
-        _.each(posts, (post, index) => {
-          const previous = index === posts.length - 1 ? null : posts[index + 1].node;
-          const next = index === 0 ? null : posts[index - 1].node;
+  // 3. Loop over each blog post and create a page
+  posts.forEach((post, index) => {
+    const previous = index === posts.length - 1 ? null : posts[index + 1].node;
+    const next = index === 0 ? null : posts[index - 1].node;
 
-          createPage({
-            path: post.node.fields.slug,
-            component: blogPost,
-            context: {
-              slug: post.node.fields.slug,
-              previous,
-              next,
-            },
-          })
-        })
-      })
-    )
-  })
+    actions.createPage({
+      path: post.node.fields.slug,
+      component: blogTemplate,
+      context: {
+        slug: post.node.fields.slug,
+        previous,
+        next,
+      },
+    });
+  });
 }
 
-exports.onCreateNode = ({ node, actions, getNode }) => {
-  const { createNodeField } = actions
+export async function createPages(params) {
+  await Promise.all([createBlogPosts(params)]);
+}
 
+export function onCreateNode({ node, actions, getNode }) {
   if (node.internal.type === `MarkdownRemark`) {
-    const value = createFilePath({ node, getNode })
-    createNodeField({
+    const value = createFilePath({ node, getNode });
+    actions.createNodeField({
       name: `slug`,
       node,
       value,
-    })
+    });
   }
 }
