@@ -1,74 +1,68 @@
-import React from 'react'
-import Helmet from 'react-helmet'
-import Link from 'gatsby-link'
-import get from 'lodash/get'
-import { graphql } from 'gatsby'
-import Layout from '../components/layout'
+import React from "react";
+import Helmet from "react-helmet";
+import Link from "gatsby-link";
+import { graphql } from "gatsby";
+import Layout from "../components/layout";
+import styled from "styled-components";
 
-import Bio from '../components/Bio'
-import { rhythm, scale } from '../utils/typography'
+import Bio from "../components/Bio";
+import { rhythm } from "../utils/typography";
 
-class BlogPostTemplate extends React.Component {
-  render() {
-    const post = this.props.data.markdownRemark
-    const siteTitle = get(this.props, 'data.site.siteMetadata.title')
-    const { previous, next } = this.props.pageContext
+export default function BlogPostTemplate(props) {
+  const post = props.data.markdownRemark;
+  const siteTitle = props.data.site.siteMetadata.title;
+  const { previous, next } = props.pageContext;
 
-    return (
-      <Layout location={this.props.location}>
-        <div>
-          <Helmet title={`${post.frontmatter.title} | ${siteTitle}`} />
-          <h1>{post.frontmatter.title}</h1>
-          <p
-            style={{
-              ...scale(-1 / 5),
-              display: 'block',
-              marginBottom: rhythm(1),
-              marginTop: rhythm(-1),
-            }}
-          >
-            {post.frontmatter.date}
-          </p>
-          <div dangerouslySetInnerHTML={{ __html: post.html }} />
-          <hr
-            style={{
-              marginBottom: rhythm(1),
-            }}
-          />
-          <Bio />
+  const DateStyle = styled.p`
+    /* ...scale(-1 / 5); */
+    display: block;
+    margin-bottom: ${rhythm(1)};
+    margin-top: ${rhythm(-1)};
+  `;
 
-          <ul
-            style={{
-              display: 'flex',
-              flexWrap: 'wrap',
-              justifyContent: 'space-between',
-              listStyle: 'none',
-              padding: 0,
-            }}
-          >
-            {previous && (
-              <li>
-                <Link to={previous.fields.slug} rel="prev">
-                  ← {previous.frontmatter.title}
-                </Link>
-              </li>
-            )}
+  const BreakStyle = styled.hr`
+    margin-bottom: ${rhythm(1)};
+  `;
 
-            {next && (
-              <li>
-                <Link to={next.fields.slug} rel="next">
-                  {next.frontmatter.title} →
-                </Link>
-              </li>
-            )}
-          </ul>
-        </div>
-      </Layout>
-    )
-  }
+  const PrevNext = styled.ul`
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    list-style: none;
+    padding: 0;
+  `;
+
+  return (
+    <Layout location={props.location}>
+      <div>
+        <Helmet title={`${post.frontmatter.title} | ${siteTitle}`} />
+        <h1>{post.frontmatter.title}</h1>
+        <DateStyle>{post.frontmatter.date}</DateStyle>
+        <div dangerouslySetInnerHTML={{ __html: post.html }} />
+        <BreakStyle />
+        <Bio />
+
+        <PrevNext>
+          {previous && (
+            <li>
+              <Link to={previous.fields.slug} rel="prev">
+                ← {previous.frontmatter.title}
+              </Link>
+            </li>
+          )}
+
+          {next && (
+            <li>
+              <Link to={next.fields.slug} rel="next">
+                {next.frontmatter.title} →
+              </Link>
+            </li>
+          )}
+        </PrevNext>
+      </div>
+    </Layout>
+  );
 }
-
-export default BlogPostTemplate
 
 export const pageQuery = graphql`
   query BlogPostBySlug($slug: String!) {
@@ -87,4 +81,4 @@ export const pageQuery = graphql`
       }
     }
   }
-`
+`;
